fix: address PR review — Steward auth hardening and correctness
- Extract shared steward-server.ts with getStewardJwtSecret(), ensureStewardUser(),
  STEWARD_API_URL, and STEWARD_PLATFORM_KEY; eliminates duplicate init code across
  farcaster/route.ts, telegram-miniapp/route.ts, and session/route.ts
- getStewardJwtSecret() throws in production when STEWARD_JWT_SECRET is missing
  (farcaster + telegram routes previously fell back silently to dev secret)
- Raise JWT expiry from 15m → 24h for Farcaster and Telegram tokens (session cookie
  is 30 days; 15m caused silent logouts during normal usage)
- Add hasFarcaster: true to new Farcaster user insert (was missing, broke social
  login detection in OnboardingProvider)
- Refresh telegramUsername (and stewardId) on every existing Telegram user login
  (username can change in Telegram; was only set on initial creation)
- Fix stale comments: middleware.ts "Privy still enabled", users/me OpenAPI PrivyAuth → StewardAuth

Made-with: Cursor
